Create ai_stock_analyzer
This is simple script that fetches financial data from yfinance that then calculates the current ratio, quick ratio, debt to equity, return on assets, gross profit margin, return on equity, earnings per share and price to earnings. That data then is sent to a gemini ai and responds with a detailed description on why you should buy, hold, or sell that stock.